feat: drinks 도메인 추가
* test: Drinks service 관련 tests 추가

* feat: domain layer 추가

* feat: application layer 추가

* feat: infra_structure layer 추가

* feat: external interface layer 추가

* feat: VO 에 filtering Enums 추가

* feat: drink review add, delete methods 추가

* test: find by drink id, create, update drink review (add, delete) 테스트

* feat: Application service 에 update, delete methods 추가

* feat: Drinks find_all with filtering 기능 추가

* feat: External interface에 POST, GET 추가

* test: drinks router tests 추가, application service test 완료

* feat: wishes service, drinks service 에 wish 추가, 삭제 기능 추가

* refactor: DrinkId VO 추가

* test: wishes test_service에 drinks 관련 기능 추가

* feat: review Create, Update, Delete시 drinks rating update기능 추가

* refactor: drinks router QueryParam 추가

* refactor: import optimization

* style: formatting with black and isort

+++ app/wishes/application/service.py
@@ -1,31 +1,33 @@
 import time
 from typing import Union
 
+from drinks.application.dtos import (AddDrinkWishInputDto,
+                                     DeleteDrinkWishInputDto)
+from drinks.application.service import DrinkApplicationService
 from shared_kernel.application.dtos import FailedOutputDto
-from wishes.application.dto import (
-    CreateWishInputDto,
-    CreateWishOutputDto,
-    DeleteWishInputDto,
-    DeleteWishOutputDto,
-    FindWishesInputDto,
-    FindWishesOutputDto,
-)
+from wishes.application.dto import (CreateWishInputDto, CreateWishOutputDto,
+                                    DeleteWishInputDto, DeleteWishOutputDto,
+                                    FindWishesInputDto, FindWishesOutputDto)
 from wishes.domain.entities import Wish
-from wishes.domain.repository import WishRepository, QueryParam
-from wishes.domain.value_objects import UserId, DrinkId
+from wishes.domain.repository import QueryParam, WishRepository
+from wishes.domain.value_objects import DrinkId, UserId
 
 
 class WishApplicationService:
     def __init__(self, wish_repository: WishRepository) -> None:
         self._wish_repository = wish_repository
 
-    def find_wishes(self, input_dto: FindWishesInputDto) -> Union[FindWishesOutputDto, FailedOutputDto]:
+    def find_wishes(
+        self, input_dto: FindWishesInputDto
+    ) -> Union[FindWishesOutputDto, FailedOutputDto]:
         try:
             wishes = self._wish_repository.find_all(input_dto.query_param)
             return FindWishesOutputDto(
                 items=[
                     FindWishesOutputDto.Item(
-                        user_id=str(wish.user_id), drink_id=str(wish.drink_id), created_at=wish.created_at
+                        user_id=str(wish.user_id),
+                        drink_id=str(wish.drink_id),
+                        created_at=wish.created_at,
                     )
                     for wish in wishes
                 ]
@@ -33,26 +35,67 @@ class WishApplicationService:
         except Exception as e:
             return FailedOutputDto.build_system_error(message=str(e))
 
-    def create_wish(self, input_dto: CreateWishInputDto) -> Union[CreateWishOutputDto, FailedOutputDto]:
+    def create_wish(
+        self,
+        input_dto: CreateWishInputDto,
+        drink_application_service: DrinkApplicationService,
+    ) -> Union[CreateWishOutputDto, FailedOutputDto]:
         try:
             user_id = UserId(value=input_dto.user_id)
             drink_id = DrinkId.from_str(input_dto.drink_id)
             wish = Wish(user_id=user_id, drink_id=drink_id, created_at=time.time())
-            if self._wish_repository.find(QueryParam(user_id=str(user_id), drink_id=str(drink_id))) is not None:
-                return FailedOutputDto.build_resource_conflict_error(f"해당 리소스는 이미 존재합니다.")
+
+            if (
+                self._wish_repository.find(
+                    QueryParam(user_id=str(user_id), drink_id=str(drink_id))
+                )
+                is not None
+            ):
+                return FailedOutputDto.build_resource_conflict_error(
+                    f"해당 리소스는 이미 존재합니다."
+                )
+
             self._wish_repository.add(wish)
-            return CreateWishOutputDto(user_id=str(user_id), drink_id=str(drink_id), created_at=wish.created_at)
+
+            input_dto = AddDrinkWishInputDto(drink_id=input_dto.drink_id)
+            add_drink_wish_output_dto = drink_application_service.add_drink_wish(
+                input_dto=input_dto
+            )
+
+            if not add_drink_wish_output_dto.status:
+                return add_drink_wish_output_dto
+            return CreateWishOutputDto(
+                user_id=str(user_id), drink_id=str(drink_id), created_at=wish.created_at
+            )
         except Exception as e:
             return FailedOutputDto.build_system_error(message=str(e))
 
-    def delete_wish(self, input_dto: DeleteWishInputDto) -> Union[DeleteWishOutputDto, FailedOutputDto]:
+    def delete_wish(
+        self,
+        input_dto: DeleteWishInputDto,
+        drink_application_service: DrinkApplicationService,
+    ) -> Union[DeleteWishOutputDto, FailedOutputDto]:
         try:
             user_id = UserId(value=input_dto.user_id)
             drink_id = DrinkId.from_str(input_dto.drink_id)
-            wish = self._wish_repository.find(QueryParam(user_id=str(user_id), drink_id=str(drink_id)))
+            wish = self._wish_repository.find(
+                QueryParam(user_id=str(user_id), drink_id=str(drink_id))
+            )
+
             if wish is None:
-                return FailedOutputDto.build_resource_not_found_error(f"해당 리소스를 찾지 못했습니다.")
+                return FailedOutputDto.build_resource_not_found_error(
+                    f"해당 리소스를 찾지 못했습니다."
+                )
+
             self._wish_repository.delete(wish)
+
+            input_dto = DeleteDrinkWishInputDto(drink_id=input_dto.drink_id)
+            delete_drink_wish_output_dto = drink_application_service.delete_drink_wish(
+                input_dto=input_dto
+            )
+
+            if not delete_drink_wish_output_dto.status:
+                return delete_drink_wish_output_dto
+            return DeleteWishOutputDto()
         except Exception as e:
             return FailedOutputDto.build_system_error(message=str(e))
-        return DeleteWishOutputDto()
